mptcp: Fix Bug #43 #27: Initialize mptcp-reqsk earlier
mtreq->mpcb is used in mptcp_reqsk_destructor to decide, whether we are
handling a new subflow or a new MPTCP-connection. Depending on this we
either remove the request-sock from the collide_tuple or the
collide_tk-list.

As mpcb was initialized so late, it might be NULL, thus trying to remove
the req-sock from the collide_tk, which is actually garbage, possibly
corrupting the list, leading to bug
https://github.com/multipath-tcp/mptcp/issues/27.

Fixes: 560443b12b731 (Don't use get_random_bytes)
Fixes: Bug #27 (https://github.com/multipath-tcp/mptcp/issues/27)
Fixes: Bug #43 (https://github.com/multipath-tcp/mptcp/issues/43)
Reported-by: chenjianqing <https://github.com/cjqing0925>
Reported-by: <liuyun_piaoye@163.com>
Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>